* Allow GL_DEPTH_COMPONENTxxx texture creation * Avoid recursion in Shutdown safe wrappers * Added gl.GetMatrixData()
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@3785 37977431-3df6-0310-b722-df95706aa16b

+++ rts/Lua/LuaShaders.cpp
@@ -326,7 +326,7 @@ int CLuaShaders::UseShader(lua_State* L)
 {
 	CheckDrawingEnabled(L, __FUNCTION__);
 
-	const int progID = luaL_checknumber(L, 1);
+	const int progID = (int)luaL_checknumber(L, 1);
 	if (progID == 0) {
 		glUseProgram(0);
 		lua_pushboolean(L, true);
